Improve handling of RSA keys and update libraries (#21)
This PR improves the way RSA keys are loaded and handled; the key is loaded once into an object when constructing the RsaPrivateKey and it's validated at the same time. When it's used the already loaded version is used, instead of letting the JWT library once more parse the PEM format and load it each time it's used.

Updated a bit of libraries and also marked cryptography as a direct dependency of the project due to the fact we use it for the handling of the keys. It's been earlier pulled in just by the other libraries using it.